Scripts, Material/Pipelines Editor Fixes method calls for GameObject functions with optional parameters
Fixes the uniforms/shader-variables editor of the Material Editor and Pipeline Editor

Slight restructuring to help with exception catching

Fixed NULL-deref when the Engine shuts down without ever loading a Luau script

Improved ref count system for GameObjects

Fixes Sounds not being paused when they are disabled

+++ src/decl/datatype/GameObject.hpp
@@ -196,10 +196,7 @@ private:
 
 	static inline Reflection::Api s_Api{};
 
-	// i think something fishy would have to be going on
-	// for the refcount to be >255
-	// 24/12/2024
-	uint8_t m_HardRefCount = 0;
+	uint16_t m_HardRefCount = 0;
 };
 
 struct GameObjectRef
